requirements.txt
python-jose
passlib
python-slugify
sqlalchemy
pydantic
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pydantic[email]